backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
neo4j==5.27.0
openai==1.59.3
beautifulsoup4==4.12.3
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
